+++ src/params.ts
@@ -11,6 +11,8 @@ export interface CancelOrder {
 
 export interface ClosePosition {
   symbol: string
+  qty?: number
+  percentage?: number
 }
 
 export interface CreateWatchList {